+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.0-dev
 
 - Enhancements
+    - Log entries for write requests (`%Instream.Log.WriteEntry{}`) now contain the actual HTTP response status returned from the writer
     - Support for deleting points using the InfluxDB v2.x "delete" (predicate) API ([#75](https://github.com/mneudert/instream/pull/75))
 
 ## 2.0.0 (2022-06-18)